Split job show viewing tests into 3 files based on role
